% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci}
\alias{ci}
\title{Confidence interval.}
\usage{
ci(x, alpha = 0.05, ignore_df0 = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{The alpha value for the confidence level.}

\item{ignore_df0}{Ignore warnings if the degrees of freedom are 0.}
}
\description{
Calculate the confidence interval of a vector.
}
