% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv}
\alias{cv}
\title{Confidence value.}
\usage{
cv(x, alpha = 0.05, ignore_df0 = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{alpha}{The alpha value for the confidence level.}

\item{ignore_df0}{Ignore warnings if the degrees of freedom are 0.}
}
\description{
Calculate the t-distribution confidence value of a vector.
}
