% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdsim.R
\docType{package}
\name{stdsim}
\alias{stdsim}
\alias{stdsim-package}
\title{stdsim: simulate standard and sample measurement for measurement optimisation}
\description{
stdsim: simulate standard and sample measurement for measurement optimisation
}
\section{Package options}{

The functions in this package listen to the \code{genplot} and
\code{verbose} options.
}

\author{
Ilja J. Kocken, \email{i.j.kocken@uu.nl}

This package incorporates functions to optimize the distribution of
(carbonate) standards and the optimal proportion of standards versus the
number of sample replicates.
}
